+++ src/Expanders.jl
@@ -529,15 +529,32 @@ end
 # @example
 # --------
 
+# Find if there is any format with color output
+function _any_color_fmt(doc)
+    idx = findfirst(x -> x isa Documenter.HTML, doc.user.format)
+    idx === nothing && return false
+    return doc.user.format[idx].ansicolor
+end
+
 function Selectors.runner(::Type{ExampleBlocks}, x, page, doc)
+    matched = match(r"^@example(?:\s+([^\s;]+))?\s*(;.*)?$", x.language)
+    matched === nothing && error("invalid '@example' syntax: $(x.language)")
+    name, kwargs = matched.captures
     # The sandboxed module -- either a new one or a cached one from this page.
-    name = match(r"^@example[ ]?(.*)$", first(split(x.language, ';', limit = 2)))[1]
     mod = Utilities.get_sandbox_module!(page.globals.meta, "atexample", name)
     sym = nameof(mod)
     lines = Utilities.find_block_in_file(x.code, page.source)
 
-    # "parse" keyword arguments to example (we only need to look for continued = true)
-    continued = occursin(r"continued\s*=\s*true", x.language)
+    # "parse" keyword arguments to example
+    continued = false
+    ansicolor = _any_color_fmt(doc)
+    if kwargs !== nothing
+        continued = occursin(r"\bcontinued\s*=\s*true\b", kwargs)
+        matched = match(r"\bansicolor\s*=\s*(true|false)\b", kwargs)
+        if matched !== nothing
+            ansicolor = matched[1] == "true"
+        end
+    end
 
     # Evaluate the code block. We redirect stdout/stderr to `buffer`.
     result, buffer = nothing, IOBuffer()
@@ -550,7 +567,7 @@ function Selectors.runner(::Type{ExampleBlocks}, x, page, doc)
             code = x.code
         end
         for (ex, str) in Utilities.parseblock(code, doc, page; keywords = false)
-            c = IOCapture.capture(rethrow = InterruptException) do
+            c = IOCapture.capture(rethrow = InterruptException, color = ansicolor) do
                 cd(page.workdir) do
                     Core.eval(mod, ex)
                 end
@@ -579,10 +596,10 @@ function Selectors.runner(::Type{ExampleBlocks}, x, page, doc)
     input   = droplines(x.code)
 
     # Generate different  in different formats and let each writer select
-    output = Base.invokelatest(Utilities.display_dict, result)
+    output = Base.invokelatest(Utilities.display_dict, result, context = :color => ansicolor)
 
     # Only add content when there's actually something to add.
-    isempty(input)  || push!(content, Markdown.Code("julia", input))
+    isempty(input) || push!(content, Markdown.Code("julia", input))
     if result === nothing
         code = Documenter.DocTests.sanitise(buffer)
         isempty(code) || push!(content, Dict{MIME,Any}(MIME"text/plain"() => code))
@@ -597,10 +614,21 @@ end
 # -----
 
 function Selectors.runner(::Type{REPLBlocks}, x, page, doc)
-    matched = match(r"^@repl[ ]?(.*)$", x.language)
+    matched = match(r"^@repl(?:\s+([^\s;]+))?\s*(;.*)?$", x.language)
     matched === nothing && error("invalid '@repl' syntax: $(x.language)")
-    name = matched[1]
+    name, kwargs = matched.captures
+    # The sandboxed module -- either a new one or a cached one from this page.
     mod = Utilities.get_sandbox_module!(page.globals.meta, "atexample", name)
+
+    # "parse" keyword arguments to repl
+    ansicolor = _any_color_fmt(doc)
+    if kwargs !== nothing
+        matched = match(r"\bansicolor\s*=\s*(true|false)\b", kwargs)
+        if matched !== nothing
+            ansicolor = matched[1] == "true"
+        end
+    end
+
     code = split(x.code, '\n'; limit = 2)[end]
     result, out = nothing, IOBuffer()
     for (ex, str) in Utilities.parseblock(x.code, doc, page; keywords = false)
@@ -611,7 +639,7 @@ function Selectors.runner(::Type{REPLBlocks}, x, page, doc)
             # see https://github.com/JuliaLang/julia/pull/33864
             ex = REPL.softscope!(ex)
         end
-        c = IOCapture.capture(rethrow = InterruptException) do
+        c = IOCapture.capture(rethrow = InterruptException, color = ansicolor) do
             cd(page.workdir) do
                 Core.eval(mod, ex)
             end
@@ -639,10 +667,10 @@ end
 # ------
 
 function Selectors.runner(::Type{SetupBlocks}, x, page, doc)
-    matched = match(r"^@setup[ ](.+)$", x.language)
+    matched = match(r"^@setup(?:\s+([^\s;]+))?\s*$", x.language)
     matched === nothing && error("invalid '@setup <name>' syntax: $(x.language)")
-    # The sandboxed module -- either a new one or a cached one from this page.
     name = matched[1]
+    # The sandboxed module -- either a new one or a cached one from this page.
     mod = Utilities.get_sandbox_module!(page.globals.meta, "atexample", name)
 
     # Evaluate whole @setup block at once instead of piecewise